feat(ui): added CameraXActivity
CameraXの基本的な機能を確認するためのアクティビティを作成した
