+++ camera/app/src/main/java/com/example/camera/camerax/CameraXActivity.kt
@@ -1,0 +1,96 @@
+package com.example.camera.camerax
+
+import android.Manifest
+import android.content.pm.PackageManager
+import android.os.Build
+import android.os.Bundle
+import android.widget.Toast
+import androidx.annotation.RequiresApi
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
+import androidx.core.content.ContextCompat
+import com.example.camera.R
+import com.example.camera.databinding.ActivityCameraXBasicBinding
+import kotlinx.android.synthetic.main.activity_camera_x_basic.*
+import kotlinx.android.synthetic.main.activity_main.*
+import java.io.File
+import java.util.concurrent.ExecutorService
+import java.util.concurrent.Executors
+
+class CameraXActivity : AppCompatActivity() {
+
+    private val camera: CameraFeature = CameraFeature()
+
+    private lateinit var outputDirectory: File
+    private lateinit var cameraExecutor: ExecutorService
+
+    private lateinit var binding: ActivityCameraXBasicBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityCameraXBasicBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        binding.cameraCaptureButton.setOnClickListener { takePhoto() }
+
+        // カメラの許可をもらう
+        if (isPermissionGranted()) startCamera()
+        else ActivityCompat.requestPermissions(
+            this,
+            arrayOf(Manifest.permission.CAMERA),
+            REQUEST_CODE_PERMISSIONS
+        )
+
+        outputDirectory = getOutputDirectory()
+        cameraExecutor = Executors.newSingleThreadExecutor()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        cameraExecutor.shutdown()
+    }
+
+    override fun onRequestPermissionsResult(
+        requestCode: Int,
+        permissions: Array<String>,
+        grantResults: IntArray
+    ) {
+        if (requestCode == REQUEST_CODE_PERMISSIONS) {
+            if (isPermissionGranted()) {
+                startCamera()
+            } else {
+                Toast.makeText(this, "許可が与えられていません.", Toast.LENGTH_SHORT).show()
+                finish()
+            }
+        }
+    }
+
+    private fun startCamera() {
+        camera.startCamera(this, this, viewFinder)
+    }
+
+    private fun takePhoto() {
+        camera.takePhoto(this, baseContext, outputDirectory)
+    }
+
+    private fun isPermissionGranted(): Boolean {
+        val permission = ContextCompat.checkSelfPermission(
+            baseContext,
+            Manifest.permission.CAMERA
+        )
+        return permission == PackageManager.PERMISSION_GRANTED
+    }
+
+    @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
+    private fun getOutputDirectory(): File {
+        val mediaDir = externalMediaDirs.firstOrNull()?.let {
+            File(it, resources.getString(R.string.app_name)).apply { mkdirs() }
+        }
+        return if (mediaDir != null && mediaDir.exists())
+            mediaDir else filesDir
+    }
+
+    companion object {
+        private const val REQUEST_CODE_PERMISSIONS = 10
+    }
+}
