+++ camera/app/src/main/res/values/strings.xml
@@ -1,3 +1,4 @@
 <resources>
     <string name="app_name">Camera</string>
+    <string name="camerax_basics">CameraX Basics</string>
 </resources>
